Extract anchor verification logic
